cheshire-backend/requirements.txt
accelerate==1.12.0
agent-framework==1.0.0b251120
agent-framework-core==1.0.0b251001
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
av==16.1.0
azure-core==1.38.2
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.26.0b2
azure-monitor-opentelemetry==1.8.6
azure-monitor-opentelemetry-exporter==1.0.0b48
banks==2.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.1
dirtyjson==1.0.8
distro==1.9.0
docling==2.74.0
docling-core==2.65.2
docling-ibm-models==3.11.0
docling-parse==5.3.4
et_xmlfile==2.0.0
Faker==40.4.0
fastapi==0.128.5
filelock==3.20.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
grpcio==1.78.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msal==1.35.0b1
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
ollama==0.6.1
omegaconf==2.3.0
openai==2.21.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-asgi==0.60b0
opentelemetry-instrumentation-dbapi==0.60b0
opentelemetry-instrumentation-django==0.60b0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-instrumentation-flask==0.60b0
opentelemetry-instrumentation-psycopg2==0.60b0
opentelemetry-instrumentation-requests==0.60b0
opentelemetry-instrumentation-urllib==0.60b0
opentelemetry-instrumentation-urllib3==0.60b0
opentelemetry-instrumentation-wsgi==0.60b0
opentelemetry-proto==1.39.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
polyfactory==3.3.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pylatexenc==2.10
pypdfium2==5.5.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
rapidocr==3.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scipy==1.17.1
semchunk==2.2.2
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.6.0
typer==0.21.2
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
wrapt==1.17.3
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
